rename processing DbSession to DatabaseSession
