+++ backend/corpora/lambdas/api/v1/submission.py
@@ -1,6 +1,6 @@
 from flask import make_response, jsonify
 
-from ....common.utils.db_session import db_session
+from ....common.utils.database_session import db_session
 from ....common.utils.exceptions import ForbiddenHTTPException
 from ....common.entities import Collection
