+++ tests/unit/backend/corpora/common/utils/test_db_utils.py
@@ -1,5 +1,5 @@
 from backend.corpora.common.corpora_orm import DbDataset
-from backend.corpora.common.utils.db_session import DbSession, db_session_manager
+from backend.corpora.common.utils.database_session import DatabaseSession, db_session_manager
 from backend.corpora.common.utils.exceptions import CorporaException
 from tests.unit.backend.fixtures.data_portal_test_case import DataPortalTestCase
 
@@ -7,8 +7,8 @@ from tests.unit.backend.fixtures.data_portal_test_case import DataPortalTestCase
 class TestDbUtils(DataPortalTestCase):
     def test__single_session(self):
         """Test that only a single sessions is created"""
-        db1 = DbSession()
-        db2 = DbSession()
+        db1 = DatabaseSession()
+        db2 = DatabaseSession()
 
         self.assertEqual(db1.session, db2.session)
